ai/face_rec_system/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
pillow==10.1.0
face-recognition==1.3.0